Defines initial read_iceberg function
